requirements.txt
openai
faiss-cpu
streamlit
python-dotenv
requests
meteostat
pandas
matplotlib
textblob